move Test class to separate file
