+++ src/main/java/FormulaSimplifier/Test.java
@@ -8,9 +8,6 @@ import java.util.List;
 public class Test {
 
     List<Condition> ors;
-    String testString;
-   // String question;
-   // Response response;
     Condition currentCondition;
 
     public Test(List<Condition> ors) {
@@ -18,7 +15,6 @@ public class Test {
     }
 
     public Test (String testString) {
-        this.testString = testString;
         this.ors = separateSiblings(testString, null, "||");
         this.ors.sort((Condition a1, Condition a2) -> {
             return a1.complexity() - a2.complexity();
@@ -26,13 +22,12 @@ public class Test {
         this.ors.forEach(o -> {
             o.siblings = this.ors;
         });
-            this.currentCondition = getMostBasicQuestion();
-
+        this.currentCondition = getMostBasicQuestion();
     }
 
     public Response answerQuestion(boolean answer, Condition currentCondition) {
 
-        if (answer && currentCondition instanceof Or) {
+        if (answer && currentCondition instanceof Condition.Or) {
             if (currentCondition.parent != null) {
                 int parentIndex = currentCondition.parent.siblings.indexOf(currentCondition.parent);
                 List<Condition> parentSiblings = currentCondition.parent.siblings;
@@ -43,7 +38,7 @@ public class Test {
             return new Response(true);
         }
 
-        if (!answer && currentCondition instanceof Or) {
+        if (!answer && currentCondition instanceof Condition.Or) {
             int currentConditionIndex = currentCondition.siblings.indexOf(currentCondition);
             if (currentConditionIndex + 1 < currentCondition.siblings.size()) {
                 return new Response(currentCondition.siblings.get(currentConditionIndex + 1).getDeepestDescendant());
@@ -51,7 +46,7 @@ public class Test {
             return new Response(false);
         }
 
-        if (answer && currentCondition instanceof And) {
+        if (answer && currentCondition instanceof Condition.And) {
             int currentConditionIndex = currentCondition.siblings.indexOf(currentCondition);
             if (currentConditionIndex + 1 < currentCondition.siblings.size()) {
                 return new Response(currentCondition.siblings.get(currentConditionIndex + 1).getDeepestDescendant());
@@ -59,7 +54,7 @@ public class Test {
             return new Response(true);
         }
 
-        if (!answer && currentCondition instanceof And) {
+        if (!answer && currentCondition instanceof Condition.And) {
             int parentIndex = currentCondition.parent.siblings.indexOf(currentCondition.parent);
             List<Condition> parentSiblings = currentCondition.parent.siblings;
             if (parentIndex + 1 < parentSiblings.size()) {
@@ -69,14 +64,6 @@ public class Test {
         }
 
         return null;
-    }
-
-    public Condition firstQuestion() {
-        return getMostBasicQuestion();
-    }
-
-    public String nextQuestion() {
-        return "nextQuestion ?";
     }
 
     private Condition getMostBasicQuestion() {
@@ -97,10 +84,10 @@ public class Test {
                 String segment = remainingString.substring(prevOrIndex, orInd);
                 if (!isWithinParentheses(segment)) {
                     if (separator.equals("&&")) {
-                        output.add(new And(trimAndStripParentheses(segment)));
+                        output.add(new Condition.And(trimAndStripParentheses(segment)));
                     }
                     else {
-                        output.add(new Or(trimAndStripParentheses(segment)));
+                        output.add(new Condition.Or(trimAndStripParentheses(segment)));
                     }
                     prevOrIndex = orInd + 2;
                 }
@@ -112,22 +99,22 @@ public class Test {
                 condition.parent = parent;
                 condition.siblings = output;
                 if (condition.conditionString.indexOf(otherSeparator) > -1) {
-                    if (condition instanceof And) {
-                        ((And) condition).ors = separateSiblings(condition.conditionString, condition, otherSeparator);
+                    if (condition instanceof Condition.And) {
+                        ((Condition.And) condition).ors = separateSiblings(condition.conditionString, condition, otherSeparator);
                     } else {
-                        ((Or) condition).ands = separateSiblings(condition.conditionString, condition, otherSeparator);
+                        ((Condition.Or) condition).ands = separateSiblings(condition.conditionString, condition, otherSeparator);
                     }
                 }
             });
         }
-     return output;
+        return output;
     }
 
     private static void addToOutput(List<Condition> output, String conditionString, String separator) {
         if (separator.equals("&&")) {
-            output.add(new And(conditionString));
+            output.add(new Condition.And(conditionString));
         } else {
-            output.add(new Or(conditionString));
+            output.add(new Condition.Or(conditionString));
         }
     }
 
@@ -143,62 +130,7 @@ public class Test {
         return StringUtils.countMatches(string, "(") > StringUtils.countMatches(string, ")");
     }
 
-    public static class Condition {
-        public String conditionString;
-        public Condition parent;
-        public List<Condition> siblings;
-
-            //TODO make this into a better representation of complexity
-           public int complexity(){
-               return this.conditionString.length();
-           }
-
-           public List<Condition> children() {
-               if (this instanceof Or) {return ((Or) this).ands;}
-               else return ((And) this).ors;
-           }
-
-           public boolean hasChildren() {
-               return this.children() != null;
-           }
-
-           public Condition getDeepestDescendant() {
-               Condition condition = this;
-               while (condition.hasChildren()) {
-                   condition = condition.children().get(0);
-               }
-               return condition;
-           }
-    }
-
-    public static class Or extends Condition {
-        private List<Condition> ands;
-
-        public Or(String condition) {
-            this.conditionString = condition;
-        }
-
-        public Or(List<Condition> ands) {
-            this.ands = ands;
-        }
-
-    }
-
-    public static class And extends Condition {
-        private List<Condition> ors;
-      //  private String conditionString;
-
-        public And(List<Condition> ors) {
-            this.ors = ors;
-        }
-
-        public And(String conditionString) {
-            this.conditionString = conditionString;
-        }
-    }
-
     public static class Response {
-        boolean resolved;
         Boolean resolvedOutcome;
         Condition newQuestion;
 
@@ -207,9 +139,7 @@ public class Test {
         }
 
         public Response(boolean resolvedOutcome) {
-          //  this.resolved = resolved;
             this.resolvedOutcome = resolvedOutcome;
-          //  this.newQuestion = newQuestion;
         }
     }
 }
